Seam finding optimizations

+++ include/seam.hpp
@@ -5,13 +5,15 @@ using namespace cimg_library;
 
 class Seam {
     float total_energy;
-    vector<int> positions;
+    vector<unsigned short int> positions;
 
   public:
     Seam(int height);
     friend bool operator<(const Seam &s1, const Seam &s2);
     float getEnergy() { return total_energy; };
-    vector<int> &getPositions() { return positions; };
-    void update(int col, float energy);
+    vector<unsigned short int> &getPositions() {
+      return positions;
+    };
+    void update(unsigned short int col, float energy);
     CImg<unsigned char> removeFrom(CImg<unsigned char> &img);
 };
