+++ src/seam.cpp
@@ -9,7 +9,7 @@ bool operator<(const Seam &s1, const Seam &s2) {
     return s1.total_energy < s2.total_energy;
 }
 
-void Seam::update(int col, float energy) {
+void Seam::update(unsigned short int col, float energy) {
     total_energy += energy;
     positions.push_back(col);
 }
